+++ src/main/java/com/graphaware/nlp/dsl/procedure/AnnotateProcedure.java
@@ -39,7 +39,7 @@ public class AnnotateProcedure extends AbstractDSL {
             Node result = getNLPManager().annotateTextAndPersist(request);
             return Stream.of(new NodeResult(result));
         } catch (Exception e) {
-            throw e;
+            throw new RuntimeException(e);
         }
     }